requirements.txt
asttokens==3.0.0
colorama==0.4.6
decorator==5.2.1
executing==2.2.0
filelock==3.13.1
fsspec==2024.6.1
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.4
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
munkres==1.1.4
networkx==3.3
parso==0.8.4
prompt_toolkit==3.0.51
pure_eval==0.2.3
pygame-ce==2.5.5
Pygments==2.19.1
PySide6==6.9.1
shiboken6==6.9.1
stack-data==0.6.3
sympy==1.13.3
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
